change name and move to another scope
