+++ .bitmap
@@ -561,6 +561,12 @@
         "mainFile": "index.ts",
         "rootDir": "scopes/pipelines/modules/merge-component-results"
     },
+    "modules/packages-excluder": {
+        "scope": "",
+        "version": "",
+        "mainFile": "index.ts",
+        "rootDir": "scopes/dependencies/modules/packages-excluder"
+    },
     "modules/render-template": {
         "scope": "teambit.html",
         "version": "0.0.40",
@@ -698,12 +704,6 @@
         "version": "0.0.496",
         "mainFile": "index.ts",
         "rootDir": "scopes/ui-foundation/react-router/react-router"
-    },
-    "regex/node-modules-exclude-packages": {
-        "scope": "",
-        "version": "",
-        "mainFile": "index.ts",
-        "rootDir": "scopes/toolbox/regex/node-modules-exclude-packages"
     },
     "schema": {
         "scope": "teambit.semantics",
