+++ scopes/dependencies/modules/packages-excluder/index.ts
@@ -1,0 +1,1 @@
+export { generateNodeModulesPatterns } from './generate-node-modules-patterns';
