+++ scopes/dependencies/modules/packages-excluder/packages-excluder.composition.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { nodeModulesExcludePackages } from './node-modules-exclude-packages';
+import { generateNodeModulesPatterns } from './generate-node-modules-patterns';
 
 export function LiveExample() {
   const [text, setText] = useState('react,@myorg,some-lib');
@@ -10,7 +10,7 @@ export function LiveExample() {
       <input value={text} onChange={(e) => setText(e.target.value)} style={{ width: 500 }} />
       <br />
       <div style={{ backgroundColor: '#ededed', padding: 8 }}>
-        {text && nodeModulesExcludePackages({ packages: text.split(',') })}
+        {text && generateNodeModulesPatterns({ packages: text.split(',') })}
       </div>
     </div>
   );
