+++ scopes/dependencies/modules/packages-excluder/packages-excluder.docs.md
@@ -1,9 +1,9 @@
 ---
-labels: ['typescript', 'utils', 'packages', 'node modules', 'node', 'regex']
+labels: ['typescript', 'utils', 'packages', 'node modules', 'node', 'regex', 'excluder']
 description: 'Create node modules regex with packages.'
 ---
 
-import { nodeModulesExcludePackages } from './node-modules-exclude-packages';
+import { packagesExcluder } from './packages-excluder';
 
 A function that receive an array of packages name to catch in node modules and return a regex of it.  
 The returned regex can be used in Jest `transformIgnorePatterns` to ignore specific packages.
@@ -13,6 +13,6 @@ For example:
 ```js live
 () => {
   const packagesToTransform = ['react', '@myorg', 'testing-library__dom'];
-  return nodeModulesExcludePackages({ packages: packagesToTransform });
+  return packagesExcluder({ packages: packagesToTransform });
 };
 ```
