--- scopes/toolbox/regex/node-modules-exclude-packages/index.ts
@@ -1,1 +1,0 @@
-export { nodeModulesExcludePackages } from './node-modules-exclude-packages';
